refac: Extract shared bin symlink/wrapper code to common.rs
Addresses PR feedback to deduplicate code between fleet.rs and
install.rs. Shared functions moved to common.rs:
- create_bin_symlinks
- install_wrapper_binary
- create_bin_symlink (unix)
- create_bin_shim (windows)
- update_shims_cfg (windows)
- write_conda_config
- write_frozen_marker
